proj2/backend/requirements.txt
Flask
supabase
python-dotenv
flask-cors
openai